requirements.txt
opencv-python
pyzbar
pyzxing
Pillow
numpy
